+++ static/css/style.css
@@ -587,3 +587,5 @@ main {
 .permissions-table th button:hover {
     background-color: #0056b3;
 }
+
+
