+++ templates/settings/audit_log.html
@@ -96,5 +96,9 @@
                 <a href="?page={{ page_obj.paginator.num_pages }}" class="btn btn-page">Последняя</a>
             {% endif %}
         </div>
+
+                <div class="all-btn">
+                <a href="{% url 'settings_info' %}" class="btn btn-cancel">Назад</a>
+            </div>
     </div>
 {% endblock %}
